Rename PHAL_Spatial to PHAL_Field_Source
