+++ src/CMakeLists.txt
@@ -424,7 +424,7 @@ list (APPEND SOURCES
   evaluators/utility/PHAL_MapToPhysicalFrame.cpp
   evaluators/utility/PHAL_MapToPhysicalFrameSide.cpp
   evaluators/utility/PHAL_Source.cpp
-  evaluators/utility/PHAL_Spatial.cpp
+  evaluators/utility/PHAL_Field_Source.cpp
   )
 
 if(ALBANY_STK_EXPR_EVAL)
@@ -559,8 +559,8 @@ list (APPEND HEADERS
   evaluators/utility/PHAL_MapToPhysicalFrame_Def.hpp
   evaluators/utility/PHAL_Source.hpp
   evaluators/utility/PHAL_Source_Def.hpp
-  evaluators/utility/PHAL_Spatial.hpp
-  evaluators/utility/PHAL_Spatial_Def.hpp
+  evaluators/utility/PHAL_Field_Source.hpp
+  evaluators/utility/PHAL_Field_Source_Def.hpp
   )
 
 if(ALBANY_STK_EXPR_EVAL)
